backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL
# For now, we'll use SQLite which doesn't need this
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
httpx==0.25.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
yfinance==0.2.32
ta==0.11.0
nltk==3.8.1
textblob==0.17.1
websockets==12.0
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3
beautifulsoup4==4.12.2
feedparser==6.0.10

